+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Fix time entries with very long descriptions overflowing the rendered list report block (issue #43)
 - Fix explicity `GROUP BY DATE` expression causing exceptions. (issue #45)
+- Add better error handling while rendering reports (issue #45)
 
 ## [0.5.0]